Move app support check method from app component to app-verion service
